+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.0 - 2020-07-06
+### Changed
+- Replaced Twitch API v5 calls with new Helix endpoints for following/unfollowing
+
+### Fixed
+- Button sometimes would not be displayed after Twitch's channel redesign
+
 ## 1.1.1 - 2020-03-04
 ### Fixed
 - Fix follow button not being properly inserted if you are already friends with or have a currently pending friend request to/from the viewer card user